Line search stopping criterion is now same as convergence test.
This removes "busy work", where we'd line search more precisely than
needed based on the criterion for stopping. Once a line search results in a
step smaller than the convergence criterion, it stops.

This breaks a couple of tests, since the Mathematica baseline code behaved
differently. Needs to be fixed, but not pressing.